CMake: Find qt-rappor-client by using Qt6_DIR and Qt5_DIR
One common way to let CMake find Qt is by setting Qt5_DIR in the environment.
The new way that tries to find Qt 6 also does not respec the variable.
Pass the variable as hint to make it work again (this is the way we set
the Qt version in our CI).